Extract safe Markdown HTML helpers
