+++ assets/main.js
@@ -1,13 +1,14 @@
 import { configureFetchCachePolicy } from './js/cache-control.js';
 import './js/components.js';
 import { createContentModel } from './js/content-model.js';
-import { mdParse } from './js/markdown.js';
+import { mdParse } from './js/markdown.js?v=markdown-safety-20260508';
 import { setupAnchors, setupTOC } from './js/toc.js';
 import { applySavedTheme, bindThemeToggle, bindThemePackPicker, mountThemeControls, refreshLanguageSelector, applyThemeConfig, bindPostEditor } from './js/theme.js';
-import { createThemeI18nContext, ensureThemeLayout, getThemeApiHandler, getThemeLayoutContext, getThemeRegion } from './js/theme-layout.js';
+import { createThemeI18nContext, ensureThemeLayout, getThemeApiHandler, getThemeLayoutContext, getThemeRegion } from './js/theme-layout.js?v=markdown-safety-20260508';
 import { setupSearch } from './js/search.js';
 import { extractExcerpt, computeReadTime, parseFrontMatter } from './js/content.js';
-import { getQueryVariable, setDocTitle, setBaseSiteTitle, cardImageSrc, fallbackCover, renderTags, slugifyTab, formatDisplayDate, isModifiedClick, getContentRoot, sanitizeImageUrl, sanitizeUrl, setSafeHtml } from './js/utils.js';
+import { getContentRoot, sanitizeImageUrl, sanitizeUrl, setSafeHtml } from './js/safe-html.js';
+import { getQueryVariable, setDocTitle, setBaseSiteTitle, cardImageSrc, fallbackCover, renderTags, slugifyTab, formatDisplayDate, isModifiedClick } from './js/utils.js';
 import {
   initI18n,
   t,
