+++ index.html
@@ -18,6 +18,6 @@
   <link rel="stylesheet" id="theme-pack">
 </head>
 <body>
-  <script type="module" src="assets/main.js?v=post-card-slot-safety-20260507"></script>
+  <script type="module" src="assets/main.js?v=markdown-safety-20260508"></script>
 </body>
 </html>
