+++ index_editor.html
@@ -2408,8 +2408,8 @@
     })();
   </script>
   <script type="module" src="assets/js/editor-boot.js?v=theme-manager-20260507"></script>
-  <script type="module" src="assets/js/editor-main.js?v=theme-manager-20260507"></script>
-  <script type="module" src="assets/js/composer.js?v=theme-catalog-remote-20260507"></script>
+  <script type="module" src="assets/js/editor-main.js?v=markdown-safety-20260508"></script>
+  <script type="module" src="assets/js/composer.js?v=markdown-safety-20260508"></script>
   <!-- Back to top button -->
   <button id="backToTop" class="back-to-top" aria-label="返回顶部" title="返回顶部" type="button" hidden>
     <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
